move back from pipeline package to networkpolicy
