+++ pkg/networkpolicy/baselineadminnetworkpolicy_test.go
@@ -1,4 +1,4 @@
-package pipeline
+package networkpolicy
 
 import (
 	"context"
