+++ pkg/networkpolicy/helpers.go
@@ -1,4 +1,4 @@
-package pipeline
+package networkpolicy
 
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
