+++ pkg/networkpolicy/pipeline_test.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: APACHE-2.0
 
-package pipeline
+package networkpolicy
 
 import (
 	"context"
